requirements.txt
streamlit
google-generativeai
python-dotenv
stable-diffusion
requests
